Map centerNode bug fixed Update

+++ NamelessHill-project/Assets/Script/Data/Data/PawnAgent.cs
@@ -168,7 +168,7 @@ namespace Nameless.Data
         #endregion
 
 
-        public PawnAgent(FrontPlayer frontPlayer, Slider healthBar,Area currentArea, Pawn pawn, int mapId)
+        public PawnAgent(FrontPlayer frontPlayer, Slider healthBar,Area currentArea, Pawn pawn, long mapId)
         {
             this.frontPlayer = frontPlayer;
             this.healthBar = healthBar;
